flaskr/requirements.txt
flask


google-cloud-translate

google-cloud-storage